+++ DSCResources/xPendingReboot/DSCResources/MSFT_xPendingReboot/MSFT_xPendingReboot.psm1
@@ -2,16 +2,18 @@
 {
     [CmdletBinding()]
     [OutputType([Hashtable])]
-     param
+    param
     (
-    [Parameter(Mandatory=$true)]
-    [string]$Name,
-    
-    [Parameter()]
-    [bool]$SkipCcmClientSDK
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Name,
+
+        [Parameter()]
+        [bool]
+        $SkipCcmClientSDK
     )
 
-    $ComponentBasedServicingKeys = (Get-ChildItem 'hklm:SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing\').Name
+    $ComponentBasedServicingKeys = (Get-ChildItem 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing\').Name
     if ($ComponentBasedServicingKeys)
     {
         $ComponentBasedServicing = $ComponentBasedServicingKeys.Split("\") -contains "RebootPending"
@@ -21,7 +23,7 @@
         $ComponentBasedServicing = $false
     }
 
-    $WindowsUpdateKeys = (Get-ChildItem 'hklm:SOFTWARE\Microsoft\Windows\CurrentVersion\WindowsUpdate\Auto Update\').Name
+    $WindowsUpdateKeys = (Get-ChildItem 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\WindowsUpdate\Auto Update\').Name
     if ($WindowsUpdateKeys)
     {
         $WindowsUpdate = $WindowsUpdateKeys.Split("\") -contains "RebootRequired"
@@ -31,12 +33,12 @@
         $WindowsUpdate = $false
     }
 
-    $PendingFileRename = (Get-ItemProperty 'hklm:\SYSTEM\CurrentControlSet\Control\Session Manager\').PendingFileRenameOperations.Length -gt 0
-    $ActiveComputerName = (Get-ItemProperty 'hklm:\SYSTEM\CurrentControlSet\Control\ComputerName\ActiveComputerName').ComputerName
-    $PendingComputerName = (Get-ItemProperty 'hklm:\SYSTEM\CurrentControlSet\Control\ComputerName\ComputerName').ComputerName
+    $PendingFileRename = (Get-ItemProperty 'HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\').PendingFileRenameOperations.Length -gt 0
+    $ActiveComputerName = (Get-ItemProperty 'HKLM:\SYSTEM\CurrentControlSet\Control\ComputerName\ActiveComputerName').ComputerName
+    $PendingComputerName = (Get-ItemProperty 'HKLM:\SYSTEM\CurrentControlSet\Control\ComputerName\ComputerName').ComputerName
     $PendingComputerRename = $ActiveComputerName -ne $PendingComputerName
 
-    
+
 
     if (-not $SkipCcmClientSDK)
     {
@@ -46,8 +48,9 @@
             Name='DetermineIfRebootPending'
             ErrorAction='Stop'
         }
-        
-        Try {
+
+        Try
+        {
             $CCMClientSDK = Invoke-WmiMethod @CCMSplat
         }
         Catch
@@ -71,15 +74,31 @@
 Function Set-TargetResource
 {
     [CmdletBinding()]
-     param
+    param
     (
-    [Parameter(Mandatory=$true)]
-    [string]$Name,
-    [bool]$SkipComponentBasedServicing,
-    [bool]$SkipWindowsUpdate,
-    [bool]$SkipPendingFileRename,
-    [bool]$SkipPendingComputerRename,
-    [bool]$SkipCcmClientSDK
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Name,
+
+        [Parameter()]
+        [bool]
+        $SkipComponentBasedServicing,
+
+        [Parameter()]
+        [bool]
+        $SkipWindowsUpdate,
+
+        [Parameter()]
+        [bool]
+        $SkipPendingFileRename,
+
+        [Parameter()]
+        [bool]
+        $SkipPendingComputerRename,
+
+        [Parameter()]
+        [bool]
+        $SkipCcmClientSDK
     )
 
     $global:DSCMachineStatus = 1
@@ -89,15 +108,31 @@ Function Test-TargetResource
 {
     [CmdletBinding()]
     [OutputType([Boolean])]
-     param
+    param
     (
-    [Parameter(Mandatory=$true)]
-    [string]$Name,
-    [bool]$SkipComponentBasedServicing,
-    [bool]$SkipWindowsUpdate,
-    [bool]$SkipPendingFileRename,
-    [bool]$SkipPendingComputerRename,
-    [bool]$SkipCcmClientSDK
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Name,
+
+        [Parameter()]
+        [bool]
+        $SkipComponentBasedServicing,
+
+        [Parameter()]
+        [bool]
+        $SkipWindowsUpdate,
+
+        [Parameter()]
+        [bool]
+        $SkipPendingFileRename,
+
+        [Parameter()]
+        [bool]
+        $SkipPendingComputerRename,
+
+        [Parameter()]
+        [bool]
+        $SkipCcmClientSDK
     )
 
     $status = Get-TargetResource $Name -SkipCcmClientSDK $SkipCcmClientSDK