+++ ChangeLog.md
@@ -6,6 +6,7 @@
 
 * Updates bundled xHyper-V DSC resource to v3.15.0
 * Updates bundled xPendingReboot DSC resource to v0.4.0.0
+* Fixes a bug where a certain BitLocker policy causes mounting errors
 
 ### v0.17.0 ###