+++ ChangeLog.md
@@ -7,6 +7,8 @@
 * Updates bundled xHyper-V DSC resource to v3.16.0
 * Updates bundled xPendingReboot DSC resource to v0.4.0.0
 * Fixes a bug where a certain BitLocker policy causes mounting errors
+* Switches from PowerShell.org community build server to Appveyor
+* Adds PSScriptAnalyzer linting tests
 
 ### v0.17.0 ###